jobs/crawl_internal_links/requirements.txt
google-auth==2.41.1
google-cloud-bigquery==3.38.0
